+++ build_logic/src/main/groovy/aurorasdeco/Constants.java
@@ -24,9 +24,9 @@ public final class Constants {
 
 	public static final String MINECRAFT_VERSION = "1.19.2";
 	public static final int MAPPINGS = 21;
-	public static final String LOADER_VERSION = "0.17.7";
-	public static final String QFAPI_VERSION = "4.0.0-beta.24+0.68.0";
-	public static final String QSL_VERSION = "3.0.0-beta.24";
+	public static final String LOADER_VERSION = "0.17.11";
+	public static final String QFAPI_VERSION = "4.0.0-beta.29+0.75.1";
+	public static final String QSL_VERSION = "3.0.0-beta.29";
 	public static final int JAVA_VERSION = 17;
 	// Dependencies
 	public static final String TERRAFORM_WOOD_API_VERSION = "4.1.0";