+++ src/main/java/dev/lambdaurora/aurorasdeco/registry/AurorasDecoTags.java
@@ -22,6 +22,7 @@ import net.minecraft.item.Item;
 import net.minecraft.tag.TagKey;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.world.gen.feature.StructureFeature;
 
 import static dev.lambdaurora.aurorasdeco.AurorasDeco.id;
 
@@ -50,5 +51,11 @@ public final class AurorasDecoTags {
 	public static final TagKey<Block> SMALL_LOG_PILES = TagKey.of(Registry.BLOCK_KEY, id("small_log_piles"));
 	public static final TagKey<Block> STUMPS = TagKey.of(Registry.BLOCK_KEY, id("stumps"));
 
-	public static final TagKey<Block> VEGETATION_ON_WATER_SURFACE = TagKey.of(Registry.BLOCK_KEY, new Identifier("c", "vegetation/on_water_surface"));
+	public static final TagKey<Block> VEGETATION_ON_WATER_SURFACE = TagKey.of(Registry.BLOCK_KEY,
+			new Identifier("c", "vegetation/on_water_surface")
+	);
+
+	public static final TagKey<StructureFeature> WAY_SIGN_DESTINATION_STRUCTURES = TagKey.of(Registry.STRUCTURE_WORLDGEN,
+			id("way_sign_destinations")
+	);
 }