+++ src/main/java/dev/lambdaurora/aurorasdeco/world/gen/feature/AquaticSurfacePatchFeature.java
@@ -19,11 +19,8 @@ package dev.lambdaurora.aurorasdeco.world.gen.feature;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
-import net.minecraft.block.Block;
-import net.minecraft.util.math.BlockPos;
+import dev.lambdaurora.aurorasdeco.world.gen.WorldGenUtils;
 import net.minecraft.util.math.intprovider.IntProvider;
-import net.minecraft.util.random.RandomGenerator;
-import net.minecraft.world.StructureWorldAccess;
 import net.minecraft.world.gen.feature.Feature;
 import net.minecraft.world.gen.feature.FeatureConfig;
 import net.minecraft.world.gen.feature.util.FeatureContext;
@@ -33,8 +30,8 @@ import net.minecraft.world.gen.stateprovider.BlockStateProvider;
  * Represents a feature whose goal is to place a patch of something onto a liquid surface in a way it looks like a cluster.
  *
  * @author LambdAurora
- * @version 1.0.0
- * @since 1.0.0
+ * @version 1.0.0-beta.12
+ * @since 1.0.0-beta.1
  */
 public class AquaticSurfacePatchFeature extends Feature<AquaticSurfacePatchFeature.Config> {
 	public AquaticSurfacePatchFeature(Codec<Config> codec) {
@@ -52,63 +49,19 @@ public class AquaticSurfacePatchFeature extends Feature<AquaticSurfacePatchFeatu
 		boolean success = false;
 
 		for (int circle = 0; circle < config.circleAmount(); circle++) {
-			success |= this.generateCircle(context, pos);
+			success |= WorldGenUtils.generateCircle(context.getWorld(), random, pos, config.radius().get(random),
+					config.toPlace(), config.additionFactor(), config.removalFactor());
 
-			pos.move(this.pickNextSpread(random, config.spread()), 0, this.pickNextSpread(random, config.spread()));
+			pos.move(WorldGenUtils.pickNextSpread(random, config.spread()), 0, WorldGenUtils.pickNextSpread(random, config.spread()));
 		}
 
 		return success;
 	}
 
-	/**
-	 * Generates one of the patch "circles".
-	 *
-	 * @param context the feature context
-	 * @param origin the origin of this "circle"
-	 * @return {@code true} if the circle successfully generated, or {@code false} otherwise
-	 */
-	private boolean generateCircle(FeatureContext<Config> context, BlockPos origin) {
-		var config = context.getConfig();
-		StructureWorldAccess world = context.getWorld();
-		var random = context.getRandom();
-
-		int radius = config.radius().get(random);
-		int radiusSquared = radius * radius;
-		int completeRadius = radius + 3;
-		int completeRadiusSquared = completeRadius * completeRadius;
-
-		var pos = origin.mutableCopy();
-		boolean success = false;
-
-		for (int iX = -completeRadius; iX <= completeRadius; iX++) {
-			int dZ = (int) Math.sqrt(completeRadiusSquared - iX * iX);
-
-			for (int iZ = -dZ; iZ <= dZ; iZ++) {
-				pos.set(origin.getX() + iX, origin.getY(), origin.getZ() + iZ);
-				var state = config.toPlace().getBlockState(random, pos);
-
-				if (iX * iX + iZ * iZ > radiusSquared) { // We're in the outer border, it's additions
-					if (random.nextFloat() < config.additionFactor() && state.canPlaceAt(world, pos)) {
-						world.setBlockState(pos, state, Block.NOTIFY_LISTENERS);
-						success = true;
-					}
-				} else if (random.nextFloat() > config.removalFactor() && state.canPlaceAt(world, pos)) {
-					// We're inside and we passed the removal check.
-					world.setBlockState(pos, state, Block.NOTIFY_LISTENERS);
-					success = true;
-				}
-			}
-		}
-
-		return success;
-	}
-
-	private int pickNextSpread(RandomGenerator random, int spread) {
-		return random.nextInt(spread) - random.nextInt(spread);
-	}
-
-	public record Config(BlockStateProvider toPlace, int circleAmount, IntProvider radius, int spread, int yOffset, float removalFactor, float additionFactor)
-			implements FeatureConfig {
+	public record Config(
+			BlockStateProvider toPlace, int circleAmount, IntProvider radius, int spread, int yOffset,
+			float removalFactor, float additionFactor
+	) implements FeatureConfig {
 		public static final Codec<Config> CODEC = RecordCodecBuilder.create(instance -> instance.group(
 						BlockStateProvider.TYPE_CODEC.fieldOf("to_place").forGetter(Config::toPlace),
 						Codec.INT.fieldOf("circle_amount").orElse(1).forGetter(Config::circleAmount),